Get, Update and Delete Property functions updated

+++ app/Http/Controllers/PropertyController.php
@@ -43,46 +43,68 @@ class PropertyController extends Controller
 
     public function getProperty(Request $request, $propertyId) {
         $property = Property::find($propertyId);
-        return response()->json([
-            'success' => true,
-            'message' => 'Property found successfully',
-            'data' => $property
-        ]);
+        if($property == null) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Property does not exist'
+            ]);
+        } else {
+            return response()->json([
+                'success' => true,
+                'message' => 'Property found successfully',
+                'data' => $property
+            ]);
+        }
     }
 
     public function updateProperty(Request $request, $propertyId) {
         $request->validate([
-            'name' => ['required','min:5','unique:properties,name'],
+            'name' => ['required','min:5','unique:properties,name,'. $propertyId],
             'state' => ['required'],
             'type' => ['required'],
             'bedrooms' => ['required']
         ]);
         // add updated property info to database table
             $propertyInfo = Property::find($propertyId);
-            $propertyInfo->name = $request->name;
-            $propertyInfo->slug = Str::slug($request->name);
-            $propertyInfo->state = $request->state;
-            $propertyInfo->type = $request->type;
-            $propertyInfo->bedrooms = $request->bedrooms;
-            $propertyInfo->save();
-        
-        // return succcess response
-        return response()->json([
-            'success' => true,
-            'message' => 'Property inforamtion updated successfully'
-        ]);
+            if($propertyInfo == null) {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Property does not exist'
+                ]);
+            } else {
+                $propertyInfo->name = $request->name;
+                $propertyInfo->slug = Str::slug($request->name);
+                $propertyInfo->state = $request->state;
+                $propertyInfo->type = $request->type;
+                $propertyInfo->bedrooms = $request->bedrooms;
+                $propertyInfo->save();
+            
+            // return succcess response
+            return response()->json([
+                'success' => true,
+                'message' => 'Property inforamtion updated successfully'
+            ]);
+        }
     }
 
     public function deleteProperty($propertyId) {
         $delProperty = Property::find($propertyId);
+        // Check if property exists
+        if($delProperty == null) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Property does not exist'
+            ]);
+        } else {
         // delete property
-        $delProperty->delete();
+            $delProperty->delete();
 
-        // return succcess response
-        return response()->json([
-            'success' => true,
-            'message' => 'Property deleted successfully'
-        ]);
+            // return succcess response
+            return response()->json([
+                'success' => true,
+                'message' => 'Property deleted successfully'
+            ]);
+        }
 
     }
 }
